[IMP] budget: consolidate budget_account_root and budget_analytic_account
- Move root_id field from budget_account_root to budget.account model
- Move budgetable field from budget_analytic_account to budget module
- Add account_analytic_account model extension with proper views
- Migrate Thai translations for analytic account budgetable field
- Update budget_product dependencies (remove budget_account_root)
- Update kmitl_depends_dev to remove deprecated module dependencies
- Mark deprecated modules as non-installable
- Add migration script for automatic cleanup on module upgrade

This consolidation reduces module fragmentation and simplifies dependency
management while preserving all existing functionality.
