+++ budget/__manifest__.py
@@ -1,6 +1,6 @@
 {
     "name": "KMITL Budgeting",
-    "version": "16.0.1.1.0",
+    "version": "16.0.1.2.0",
     "summary": """ KMITL Budgeting""",
     "category": "KMITL/Budgeting",
     "author": "Aginix Technologies",
@@ -28,6 +28,7 @@
         "views/budget_menus.xml",
         "views/res_config_settings_views.xml",
         "views/account_move_views.xml",
+        "views/account_analytic_account_views.xml",
     ],
     "auto_install": False,
     "application": True,
