+++ budget/models/__init__.py
@@ -20,3 +20,4 @@ from . import budget_transfer_reject_wizard
 from . import res_config_settings
 from . import budget_tree
 from . import account_move
+from . import account_analytic_account
