--- budget_account_root/__manifest__.py
@@ -1,13 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    "name": "Budget Account Root",
-    "version": "16.0.1.0.0",
-    "category": "KMITL/Budgeting",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "depends": ["budget"],
-    "data": [],
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
