--- budget_analytic_account/__manifest__.py
@@ -1,15 +1,0 @@
-# -*- coding: utf-8 -*-
-{
-    "name": "Budget Analytic Account",
-    "version": "16.0.1.0.0",
-    "category": "KMITL/Budgeting",
-    "author": "Aginix Technologies",
-    "website": "https://github.com/aginix/kmitl",
-    "depends": ["analytic"],
-    "data": [
-        "views/account_analytic_account_views.xml"
-    ],
-    "installable": True,
-    "auto_install": False,
-    "license": "LGPL-3",
-}
