--- budget_analytic_account/i18n/th.po
@@ -1,37 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* budget_analytic_account
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 16.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2025-10-27 16:19+0000\n"
-"PO-Revision-Date: 2025-10-27 16:19+0000\n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: budget_analytic_account
-#: model:ir.model,name:budget_analytic_account.model_account_analytic_account
-msgid "Analytic Account"
-msgstr "บัญชีวิเคราะห์"
-
-#. module: budget_analytic_account
-#: model:ir.model.fields,field_description:budget_analytic_account.field_account_analytic_account__budgetable
-#: model_terms:ir.ui.view,arch_db:budget_analytic_account.view_account_analytic_account_search
-msgid "Budgetable"
-msgstr "ระบุงบประมาณได้"
-
-#. module: budget_analytic_account
-#: model_terms:ir.ui.view,arch_db:budget_analytic_account.view_account_analytic_account_form
-msgid "Budgeting"
-msgstr "งบประมาณ"
-
-#. module: budget_analytic_account
-#: model:ir.model.fields,help:budget_analytic_account.field_account_analytic_account__budgetable
-msgid "ติ๊กถูกเพื่อระบุว่ารหัสค่าใช้จ่ายสามารถจัดสรรงบประมาณได้"
-msgstr ""
